Add rule: leap of P5 or bigger should change the current direction
Page 7 of Counterpoint in Composition (24 in the PDF) has 1-7a, which didn't get criticized (although it was an example for when to change direction with stepwise motion), had G3 as its first note and D4 as its second, then went to C4: a P5 leap up followed by motion in the opposite direction. There was no motion preceding the P5 leap up, no direction to change. My code allows for this kind of situation. If the P5 leap starts the cantus, it can't have changed direction. My code allows that. The "motion in opposite direction" is taken care of by the rule that says things must be preceded or followed by step in opposite direction if P4 or bigger, and P5 leaps are certainly P4 or bigger. If that rule is allowed to run on a leap from the first note to the second, anyways. So P5s and up starting a cantus firmus will reliably be followed by step in opposite direction, balancing the motion.