Attempt to add rule: do not outline dissonant melodic intervals
switching directions marks the beginning and end of a sequence of notes, if that sequence's first and last note form a dissonant melodic interval (for these purposes, tritone, any seventh, any ninth), you can't do that. of course, seconds are also dissonant melodic, but I didn't put that in because I've seen a case where seconds were left alone. I'll have to check if it's when the sequence is just 2 notes, like C4 G4 F4 A4's G4 F4, or if it's allowable when the sequence is longer.